adjust router method to router link, get stop data from service
this is a more robust way of getting the stop data. The issue was that a user had to go back to the trip detail page, to get the updated data for the stop. This way, the updated data comes straight from the service. this will also make it more reusable in other components